Move into a dynamically generated Datamodel API
