+++ src/roman_datamodels/datamodels/__init__.py
@@ -1,5 +1,6 @@
 from ._core import *  # noqa: F403
 from ._datamodels import *  # noqa: F403
+from ._mixins import *  # noqa: F403
 
 # rename rdm_open to open to match the current roman_datamodels API
 from ._utils import rdm_open as open  # noqa: F403, F401
