+++ src/roman_datamodels/datamodels/_datamodels.py
@@ -6,10 +6,10 @@ This module provides all the specific datamodels used by the Roman pipeline.
     from the schema manifest defined by RAD.
 """
 
-import numpy as np
 
 from roman_datamodels import stnode
 
+from . import _mixins
 from ._core import DataModel
 
 __all__ = []
@@ -31,164 +31,30 @@ class _DataModel(DataModel):
         __all__.append(cls.__name__)
 
 
-class MosaicModel(_DataModel):
-    _node_type = stnode.WfiMosaic
+def datamodel_names():
+    for tag in stnode._stnode.DATAMODELS_MANIFEST["tags"]:
+        schema = stnode._factories.load_schema_from_uri(tag["schema_uri"])
+        if "datamodel_name" in schema:
+            yield stnode.OBJECT_NODE_CLASSES_BY_TAG[tag["tag_uri"]], schema["datamodel_name"]
 
 
-class ImageModel(_DataModel):
-    _node_type = stnode.WfiImage
+def datamodel_factory(node_type, datamodel_name):
+    if hasattr(_mixins, mixin := f"{datamodel_name}Mixin"):
+        class_type = (_DataModel, getattr(_mixins, mixin))
+    else:
+        class_type = (_DataModel,)
+
+    cls = type(
+        datamodel_name,
+        class_type,
+        {
+            "_node_type": node_type,
+            "__module__": "roman_datamodels.datamodels",
+            "__doc__": f"Roman {datamodel_name} model",
+        },
+    )
+    globals()[datamodel_name] = cls
 
 
-class ScienceRawModel(_DataModel):
-    _node_type = stnode.WfiScienceRaw
-
-
-class MsosStackModel(_DataModel):
-    _node_type = stnode.MsosStack
-
-
-class RampModel(_DataModel):
-    _node_type = stnode.Ramp
-
-    @classmethod
-    def from_science_raw(cls, model):
-        """
-        Construct a RampModel from a ScienceRawModel
-
-        Parameters
-        ----------
-        model : ScienceRawModel or RampModel
-            The input science raw model (a RampModel will also work)
-        """
-
-        if isinstance(model, cls):
-            return model
-
-        if isinstance(model, ScienceRawModel):
-            from roman_datamodels.maker_utils import mk_ramp
-
-            instance = mk_ramp(shape=model.shape)
-
-            # Copy input_model contents into RampModel
-            for key in model:
-                # If a dictionary (like meta), overwrite entries (but keep
-                # required dummy entries that may not be in input_model)
-                if isinstance(instance[key], dict):
-                    instance[key].update(getattr(model, key))
-                elif isinstance(instance[key], np.ndarray):
-                    # Cast input ndarray as RampModel dtype
-                    instance[key] = getattr(model, key).astype(instance[key].dtype)
-                else:
-                    instance[key] = getattr(model, key)
-
-            return cls(instance)
-
-        raise ValueError("Input model must be a ScienceRawModel or RampModel")
-
-
-class RampFitOutputModel(_DataModel):
-    _node_type = stnode.RampFitOutput
-
-
-class AssociationsModel(_DataModel):
-    # Need an init to allow instantiation from a JSON file
-    _node_type = stnode.Associations
-
-    @classmethod
-    def is_association(cls, asn_data):
-        """
-        Test if an object is an association by checking for required fields
-
-        Parameters
-        ----------
-        asn_data :
-            The data to be tested.
-        """
-        return isinstance(asn_data, dict) and "asn_id" in asn_data and "asn_pool" in asn_data
-
-
-class GuidewindowModel(_DataModel):
-    _node_type = stnode.Guidewindow
-
-
-class FlatRefModel(_DataModel):
-    _node_type = stnode.FlatRef
-
-
-class DarkRefModel(_DataModel):
-    _node_type = stnode.DarkRef
-
-
-class DistortionRefModel(_DataModel):
-    _node_type = stnode.DistortionRef
-
-
-class GainRefModel(_DataModel):
-    _node_type = stnode.GainRef
-
-
-class IpcRefModel(_DataModel):
-    _node_type = stnode.IpcRef
-
-
-class LinearityRefModel(_DataModel):
-    _node_type = stnode.LinearityRef
-
-    def get_primary_array_name(self):
-        """
-        Returns the name "primary" array for this model, which
-        controls the size of other arrays that are implicitly created.
-        This is intended to be overridden in the subclasses if the
-        primary array's name is not "data".
-        """
-        return "coeffs"
-
-
-class InverseLinearityRefModel(_DataModel):
-    _node_type = stnode.InverseLinearityRef
-
-    def get_primary_array_name(self):
-        """
-        Returns the name "primary" array for this model, which
-        controls the size of other arrays that are implicitly created.
-        This is intended to be overridden in the subclasses if the
-        primary array's name is not "data".
-        """
-        return "coeffs"
-
-
-class MaskRefModel(_DataModel):
-    _node_type = stnode.MaskRef
-
-    def get_primary_array_name(self):
-        """
-        Returns the name "primary" array for this model, which
-        controls the size of other arrays that are implicitly created.
-        This is intended to be overridden in the subclasses if the
-        primary array's name is not "data".
-        """
-        return "dq"
-
-
-class PixelareaRefModel(_DataModel):
-    _node_type = stnode.PixelareaRef
-
-
-class ReadnoiseRefModel(_DataModel):
-    _node_type = stnode.ReadnoiseRef
-
-
-class SuperbiasRefModel(_DataModel):
-    _node_type = stnode.SuperbiasRef
-
-
-class SaturationRefModel(_DataModel):
-    _node_type = stnode.SaturationRef
-
-
-class WfiImgPhotomRefModel(_DataModel):
-    _node_type = stnode.WfiImgPhotomRef
-
-
-class RefpixRefModel(_DataModel):
-    _node_type = stnode.RefpixRef
+for node_type, datamodel_name in datamodel_names():
+    datamodel_factory(node_type, datamodel_name)
