+++ src/roman_datamodels/stnode/__init__.py
@@ -5,5 +5,6 @@ The STNode classes and supporting objects generated dynamically at import time
 from ._converters import *  # noqa: F403
 from ._mixins import *  # noqa: F403
 from ._node import *  # noqa: F403
+from ._registry import *  # noqa: F403
 from ._stnode import *  # noqa: F403
 from ._tagged import *  # noqa: F403
